+++ CHANGELOG.md
@@ -10,8 +10,9 @@ All Notable changes to `oauth2-phabricator` will be documented in this file
 - Nothing
 
 ### Fixed
-- #2 Escape token in `Phabricator` (thanks [shadowhand](https://github.com/shadowhand)
-- #3 Possible `E_NOTICE` in `PhabricatorResourceOwner` (thanks [shadowhand](https://github.com/shadowhand) 
+- #1 Removed superfluous code ([shadowhand](https://github.com/shadowhand))
+- #2 Escape token in `Phabricator` ([shadowhand](https://github.com/shadowhand))
+- #3 Possible `E_NOTICE` in `PhabricatorResourceOwner` ([shadowhand](https://github.com/shadowhand)) 
 
 ### Removed
 - Nothing